separate emgapi reliance from queryset
